Refactored part of the left navigation panel code into its own component
Code that deals with getting the text input and creating a new project is moved to ProjectNavigationPanel.js. Changed the skeleton of
the individual lists and tasks as well.
