+++ frontend/src/pages/DashboardPage.js
@@ -1,5 +1,6 @@
 import '../styles/Dashboard.css';
 import { useState, useEffect } from 'react';
+import ProjectNavigationPanel from '../components/ProjectNavigationPanel';
 
 export default function Dashboard() {
   // GET "data" FROM BACKEND --> will end up being something like data.projects
@@ -8,30 +9,11 @@ export default function Dashboard() {
     { id: 2, name: 'Project 2' },
   ];
   const [projects, setProjects] = useState(data); // array of all project objects
-  const [inputValue, setInputValue] = useState(''); // input value for "Add a project" text field
   const [selectedProjectId, setSelectedProjectId] = useState('home'); // project id that is currently selected
 
-  // updates inputValue to be user inputted value everytime a change is detected
-  const handleChange = (event) => {
-    setInputValue(event.target.value);
-  };
-
-  // creates a new project object and adds it to list of projects
-  const handleSubmit = (event) => {
-    event.preventDefault();
-    if (inputValue === '') return;
-    let newProjectObj = createProject(inputValue);
+  // updates array of projects
+  function handleProjectAdd(newProjectObj) {
     setProjects([...projects, newProjectObj]);
-    setInputValue('');
-  };
-
-  // creates and returns a new project object
-  function createProject(projectName) {
-    return {
-      id: Date.now().toString(),
-      name: projectName,
-      tasks: [],
-    };
   }
 
   // sends updated projects array to backend (send selectedProjectId as well?????)
@@ -66,78 +48,54 @@ export default function Dashboard() {
   return (
     <div className="outer-container-div">
       <div className="navbar-div">
-        <h1>Projects</h1>
-        <p
-          className={'home' === selectedProjectId ? 'active-project' : 'home'}
-          onClick={() => handleSelect('home')}
-        >
-          Home
-        </p>
-        <div className="project-list-div">
-          <ul>
-            {projects.map((item) => (
-              <li
-                key={item.id}
-                className={item.id === selectedProjectId ? 'active-project' : 'project-list-item'}
-                onClick={() => handleSelect(item.id)}
-              >
-                {item.name}
-              </li>
-            ))}
-          </ul>
-          <form action="" className="new-project-form" onSubmit={handleSubmit}>
-            <input
-              type="text"
-              value={inputValue}
-              onChange={handleChange}
-              placeholder="New project"
-              aria-label="New project"
-            />
-            <button type="submit" className="btn-create" aria-label="create new project">
-              +
-            </button>
-          </form>
-        </div>
+        <ProjectNavigationPanel
+          selectedProjectId={selectedProjectId}
+          handleSelect={handleSelect}
+          projects={projects}
+          handleProjectAdd={handleProjectAdd}
+        />
       </div>
-      <div className="display-div">
-        <h1>Project Name</h1>
-        <button onClick={handleProjectDelete}> delete project</button>
-        <button>+ list</button>
-        <div id="list 1">
-          <div className="list-header">
-            <h3 className="list-title">List 1</h3>
-            <p className="task-count">3 tasks remaining</p>
-          </div>
-          <div className="list-body">
-            <div className="tasks-div">
-              <div className="task">
-                <input type="checkbox" id="task-1" />
-                <label for="task-1">
-                  <span class="custom-checkbox"></span>
-                  task 1
-                </label>
-              </div>
-              <div className="task">
-                <input type="checkbox" id="task-1" />
-                <label for="task-1">
-                  <span class="custom-checkbox"></span>
-                  task 2
-                </label>
-              </div>
-              <div className="task">
-                <input type="checkbox" id="task-1" />
-                <label for="task-1">
-                  <span class="custom-checkbox"></span>
-                  task 3
-                </label>
+      <div className="outer-display-div">
+        <div className="inner-display-div">
+          <h1>Project Name</h1>
+          <button onClick={handleProjectDelete}> delete project</button>
+          <button>+ list</button>
+          <div id="list 1">
+            <div className="list-header">
+              <h3 className="list-title">List 1</h3>
+              <p className="task-count">3 tasks remaining</p>
+            </div>
+            <div className="list-body">
+              <div className="tasks-div">
+                <div className="task">
+                  <input type="checkbox" id="task-1" />
+                  <label for="task-1">
+                    <span class="custom-checkbox"></span>
+                    task 1
+                  </label>
+                </div>
+                <div className="task">
+                  <input type="checkbox" id="task-1" />
+                  <label for="task-1">
+                    <span class="custom-checkbox"></span>
+                    task 2
+                  </label>
+                </div>
+                <div className="task">
+                  <input type="checkbox" id="task-1" />
+                  <label for="task-1">
+                    <span class="custom-checkbox"></span>
+                    task 3
+                  </label>
+                </div>
               </div>
             </div>
-          </div>
-          <div className="new-task-creator">
-            <form action="">
-              <input type="text" placeholder="new task name" aria-label="new task name" />
-              <button aria-label="create new task">+</button>
-            </form>
+            <div className="new-task-creator">
+              <form action="">
+                <input type="text" placeholder="new task name" aria-label="new task name" />
+                <button aria-label="create new task">+</button>
+              </form>
+            </div>
           </div>
         </div>
       </div>
